Ensure the app refreshes settings regularly by adjusting the update intervals
Fixes a type mismatch in `HiddifyNGApplication.kt` by casting `pingFrequency` values to Long for the PingWorker request.

+++ app/src/main/java/com/hiddify/hiddifyng/HiddifyNGApplication.kt
@@ -121,8 +121,8 @@ class HiddifyNGApplication : Application() {
                 Log.i(TAG, "Scheduling ping worker every $pingFrequency minutes")
                 
                 val pingRequest = PeriodicWorkRequestBuilder<PingWorker>(
-                    pingFrequency, TimeUnit.MINUTES,
-                    pingFrequency / 2, TimeUnit.MINUTES // Flex period
+                    pingFrequency.toLong(), TimeUnit.MINUTES,
+                    (pingFrequency / 2).toLong(), TimeUnit.MINUTES // Flex period
                 )
                     .setConstraints(constraints)
                     .build()
